+++ src/components/NavBar/NavBar.jsx
@@ -1,22 +1,36 @@
+import React from "react";
 import { Link } from "react-router-dom";
 import * as userService from "../../utilities/users-service";
 
 export default function NavBar({ user, setUser }) {
-  // Add the following function
   function handleLogOut() {
-    // Delegate to the users-service
     userService.logOut();
-    // Update state will also cause a re-render
     setUser(null);
   }
+
+  function renderNavLinks() {
+    if (user && user.itinerary.length > 0) {
+      return (
+        <>
+          <Link to="/itinerary/select">My Itinerary</Link>&nbsp; | &nbsp;
+        </>
+      );
+    } else {
+      return (
+        <>
+          <Link to="/itinerary">New Itinerary</Link>&nbsp; | &nbsp;
+        </>
+      );
+    }
+  }
+
   return (
     <nav>
       <Link to="/">Home</Link>&nbsp; | &nbsp;
       <Link to="/tours">Tours</Link>&nbsp; | &nbsp;
       <Link to="/ticketing">Ticketing</Link>&nbsp; | &nbsp;
       <Link to="/exhibition">Exhibition</Link>&nbsp; | &nbsp;
-      <Link to="/itinerary">New Itinerary</Link>&nbsp; | &nbsp;
-      <Link to="/itinerary/select">My Itinerary</Link>&nbsp; | &nbsp;
+      {renderNavLinks()}
       {user ? (
         <>
           <span>Welcome, {user.name}!</span>&nbsp; | &nbsp;
@@ -24,8 +38,9 @@ export default function NavBar({ user, setUser }) {
             Log Out
           </Link>
         </>
-      ) : null}
-      {!user && <Link to="/auth">Log In</Link>}
+      ) : (
+        <Link to="/auth">Log In</Link>
+      )}
     </nav>
   );
 }
