+++ src/pages/ItineraryPage/ItinerarySelectPage.jsx
@@ -1,14 +1,22 @@
-import { Link } from "react-router-dom";
+import { Link, useNavigate } from "react-router-dom";
 
 export default function ItinerarySelectPage({ user, setUser }) {
+  const navigate = useNavigate();
+
   const fetchQuery = async () => {
     try {
-      const response = await fetch("/user/deleteItinerary", {
+      const response = await fetch(`/user/${user._id}/itinerary`, {
         method: "DELETE",
       });
-      const data = await response.json();
-      setUser(data);
+      if (response.ok) {
+        const data = await response.json();
+        setUser(data);
+        navigate("/itinerary");
+      } else {
+        throw new Error("Failed to delete itinerary");
+      }
     } catch (error) {
+      console.log(user);
       console.error(error);
     }
   };
@@ -29,9 +37,8 @@ export default function ItinerarySelectPage({ user, setUser }) {
       </section>
       <section className="section-container">
         {user.itinerary?.map((each, index) => (
-            <div key={index}>{each.exhibitionName}</div>
-          ))
-       }
+          <div key={index}>{each.exhibitionName}</div>
+        ))}
       </section>
       <section className="section-container">
         <button onClick={handleDeleteItinerary}>Delete my itinerary</button>
