+++ src/routes/users.js
@@ -7,7 +7,6 @@ router.post("/", usersCtrl.create);
 router.get("/:id", usersCtrl.listOne);
 router.patch("/", usersCtrl.patch);
 
-router.delete('/deleteItinerary', usersCtrl.deleteItinerary)
-
+router.delete('/:id/itinerary', usersCtrl.deleteItinerary);
 
 module.exports = router;
